refactor: rename draw to visualize
