+++ src/featureMatching/index.ts
@@ -1,6 +1,6 @@
 export * from './matching/bruteForceMatch';
-export * from './draw/drawKeypoints';
-export * from './draw/drawMatches';
+export * from './visualize/drawKeypoints';
+export * from './visualize/drawMatches';
 export * from './descriptors/getBriefDescriptors';
 export * from './keypoints/getFastKeypoints';
 export * from './keypoints/getIntensityCentroid';
