+++ src/featureMatching/visualize/__tests__/__image_snapshots__/draw-keypoints-test-ts-alphabet-image-with-score-coloring-1-snap.png
