+++ src/featureMatching/visualize/__tests__/drawKeypoints.test.ts
