+++ src/featureMatching/visualize/drawKeypoints.ts
